Move Markdown on net.core, create project for tests
