+++ cs/Markdow/StopSymbolDecision.cs
