+++ cs/Test_Markdown/Test_Md.cs
@@ -1,8 +1,8 @@
-using System.Globalization;
 using FluentAssertions;
+using Markdown;
 using NUnit.Framework;
 
-namespace Markdown.Tests
+namespace Test_Markdown
 {
     [TestFixture]
     public class Test_Md
