+++ cs/Test_Markdown/Test_Token.cs
@@ -1,8 +1,8 @@
-using System;
-using NUnit.Framework;
 using FluentAssertions;
+using Markdown;
+using NUnit.Framework;
 
-namespace Markdown.Tests
+namespace Test_Markdown
 {
     [TestFixture]
     public class Test_Token
